--- .kdev4/GhostFragment.kdev4
@@ -1,33 +1,0 @@
-[Buildset]
-BuildItems=@Variant(\x00\x00\x00\t\x00\x00\x00\x00\x01\x00\x00\x00\x0b\x00\x00\x00\x00\x01\x00\x00\x00\x1a\x00G\x00h\x00o\x00s\x00t\x00F\x00r\x00a\x00g\x00m\x00e\x00n\x00t)
-
-[CMake]
-Build Directory Count=1
-Current Build Directory Index-Host System=0
-
-[CMake][CMake Build Directory 0]
-Build Directory Path=/home/jacob/Projects/GhostFragment/build
-Build Type=Release
-CMake Binary=/usr/bin/cmake
-CMake Executable=/usr/bin/cmake
-Environment Profile=
-Extra Arguments=-DNWX_MODULE_DIRECTORY=/home/jacob/Projects/GhostFragment/install
-Install Directory=/home/jacob/Projects/GhostFragment/install
-Runtime=Host System
-
-[CustomDefinesAndIncludes][ProjectPath0]
-Path=.
-parseAmbiguousAsCPP=true
-parserArguments=-ferror-limit=100 -fspell-checking -Wdocumentation -Wunused-parameter -Wunreachable-code -Wall -std=c++17
-parserArgumentsC=-ferror-limit=100 -fspell-checking -Wdocumentation -Wunused-parameter -Wunreachable-code -Wall -std=c99
-parserArgumentsCuda=-ferror-limit=100 -fspell-checking -Wdocumentation -Wunused-parameter -Wunreachable-code -Wall -std=c++11
-parserArgumentsOpenCL=-ferror-limit=100 -fspell-checking -Wdocumentation -Wunused-parameter -Wunreachable-code -Wall -cl-std=CL1.1
-
-[CustomDefinesAndIncludes][ProjectPath0][Compiler]
-Name=Clang
-
-[Project]
-VersionControlSupport=kdevgit
-
-[pythonsupport]
-interpreter=/home/jacob/Environments/base/bin/python
