--- GhostFragment.kdev4
@@ -1,4 +1,0 @@
-[Project]
-CreatedFrom=CMakeLists.txt
-Manager=KDevCMakeManager
-Name=GhostFragment
