+++ src/ghostfragment/hydrocarbon/position.hpp
@@ -16,10 +16,9 @@
 
 #pragma once
 #include <array>
-#include <cmath>
 #include <vector>
 
-namespace testing {
+namespace ghostfragment::hydrocarbon {
 
 std::array<float, 3> position_carbon(const std::vector<float>& source_coords,
                                      float carbon_bond, int num,
@@ -28,4 +27,4 @@ std::array<float, 3> position_carbon(const std::vector<float>& source_coords,
 std::array<float, 3> position_hydrogen(const std::vector<float>& source_coords,
                                        int flag, int num, float hydrogen_bond,
                                        float angle_deg);
-} // namespace testing
+} // namespace ghostfragment::hydrocarbon
