--- tests/cxx/unit_tests/ghostfragment/testing/hydrocarbon/position.hpp
@@ -1,31 +1,0 @@
-/*
- * Copyright 2024 GhostFragment
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#pragma once
-#include <array>
-#include <cmath>
-#include <vector>
-
-namespace testing {
-
-std::array<float, 3> position_carbon(const std::vector<float>& source_coords,
-                                     float carbon_bond, int num,
-                                     float angle_deg);
-
-std::array<float, 3> position_hydrogen(const std::vector<float>& source_coords,
-                                       int flag, int num, float hydrogen_bond,
-                                       float angle_deg);
-} // namespace testing
